end For Game

+++ Assets/Scripts/Map/ObstecleController.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using CustomEventBus;
 
@@ -21,7 +19,7 @@ public class ObstecleController : MonoBehaviour
     void Awake()
     {
         rigidbody = GetComponent<Rigidbody2D>();
-        int randomGap = Random.Range(CONSTANTS.SCRIPT_CONSTANS.MinGapBetweenObstacle, CONSTANTS.SCRIPT_CONSTANS.MaxGapBetweenObstacle);
+        float randomGap = Random.Range(CONSTANTS.SCRIPT_CONSTANS.MinGapBetweenObstacle, CONSTANTS.SCRIPT_CONSTANS.MaxGapBetweenObstacle);
         _UpperObstacle.position = new Vector2(_UpperObstacle.position.x, _UpperObstacle.position.y + randomGap / 2); 
         _BottomObstacle.position = new Vector2(_BottomObstacle.position.x, _BottomObstacle.position.y - randomGap / 2);
     }
